Run the whole server against a real database, and fix the bug it found
Supabase's REST API is PostgREST and its keys are ordinary JWTs naming a
Postgres role, so Postgres + PostgREST + a path-rewriting proxy stands in for
it — about 150MB rather than the several gigabytes the Docker stack needs,
which this machine does not have. pnpm dev:stack brings it up.

pnpm test:api then plays four players through three complete rounds in four
cookie jars. Every assertion passes, including: only the Psychic receives the
target and everyone else is refused in every phase, the target appears nowhere
in public state before the reveal, refusals leak nothing, the Psychic cannot
touch the dial, the guessing team cannot call left/right, a locked guess
cannot be moved, a clue cannot be rewritten, the active team alternates, the
Psychic rotates, teams freeze at kickoff, and a forged cookie is rejected.
Scores are re-derived independently and matched against what the server wrote.

The bug this found: take_round_target raised 42804 at call time. target_center
is a DOMAIN over smallint and RETURNS TABLE compares types strictly, so the
reveal failed on every round. read_round_target was fine only because SELECT
INTO casts implicitly on assignment — so the Psychic could see the target but
no round could ever be scored. Nothing short of executing it would have caught
this; it is invisible to typechecking and to schema inspection. Fixed with an
explicit cast and pinned by T13.

Also made service_role's grants explicit rather than relying on Supabase
defaults, since the blanket revoke above them is easy to widen by accident.

Confirmed against real PostgREST, previously listed as Supabase-only: the
private schema returns PGRST106 — to the service role as well, which is direct
evidence the SECURITY DEFINER redesign was necessary rather than precautionary.

+++ supabase/migrations/20260920000300_secret_access.sql
@@ -60,8 +60,12 @@ security definer
 set search_path = ''
 as $$
 begin
+  -- target_center is the `dial_position` DOMAIN, not a bare smallint, and
+  -- RETURNS TABLE compares the two strictly — without this cast Postgres
+  -- raises 42804 at call time. (read_round_target avoids it only because
+  -- SELECT INTO casts implicitly on assignment.)
   return query
-    select t.target_center, t.nonce
+    select t.target_center::smallint, t.nonce
       from private.round_targets t
      where t.round_id = p_round_id;
 end;
